+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Features:
 * Support Python 3.12
+* Support [PEP 695](https://peps.python.org/pep-0695/) syntax for declaring
+  type aliases
 * Introduce Y057: Do not use `typing.ByteString` or `collections.abc.ByteString`. These
   types have unclear semantics, and are deprecated; use  `typing_extensions.Buffer` or
   a union such as `bytes | bytearray | memoryview` instead. See